+++ NEWS.md
@@ -1,4 +1,4 @@
-# nanonext 0.10.4.9001 (development)
+# nanonext 0.10.4.9002 (development)
 
 #### New Features
 
@@ -7,7 +7,9 @@
 
 #### Updates
 
-* Restricts `nextmode()` to register custom serialization and unserialization functions for external pointer objects for simplicity.
+* `nextmode()` improvements:
+  + simplified function signature to take a 'refhook' argument comprising a list of serialization / unserialization functions.
+  + registered 'refhook' functions apply to external pointer type objects only.
 * `until()` updated to be identical to `.until()`, returning FALSE instead of TRUE if the timeout has been reached.
 * `pipe_notify()` arguments 'add', 'remove' and 'flag' now default to FALSE instead of TRUE for easier selective specification of the events to signal.
 * Upgrades bundled 'libmbedtls' to v 3.5.1.